backend/requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
flask-cors==4.0.0
pymongo==4.6.0
python-dotenv==1.0.0
PyJWT==2.8.0
flask-restx==1.3.0